🚚 Move google-services functions to a subdirectory
