+++ src/functions/google-services/index.ts
@@ -1,0 +1,1 @@
+export { GoogleServicesEnable } from './enable.js';
